fail Travis build if tests fail
Currently, due to a bug in Travis-CI, the build won't fail
unless the last command in the script directive fails.

This means that if the lint or tests or coverage fail, the
Travis CI build itself won't fail, in turn giving the false
sense that everything is OK and the PR can be merged.

In order to work around this Travis CI limitation we need
to explicitly ask for the shell to exit with an error
immediately if a command exits with a non-zero status.
(set -e)